Bring in Version 1.0 beta8 (snapdragon)
this is the starting point of this new develop branch
which will contain the most recent changes and fixes for
zzmoove governor. we will start here with a complete new
history. informations and main version history of beta
and stable versions can be found in all other branches.

NOTE: Consider this branch as experimental!