cpufreq: interactive: boost to max frequency if the delta between the previous sample load and the current one is big enough
This patch is slightly inspired from an upstream patch which uses sched_load
but we don't have support for load hints from the scheduler so we have do
with a simple delta between previous load and current sample load. If the delta
is wide enough it can be a dead giver about a sudden performance requirement so
boost it up all the way.
If there are consecutive deltas and we fall into that same condition we'll just
return early and rearm the timer again, there's no need to wait for the min_sample_time
since we most likely just need those sudden bursts.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: Srijith Bharadwaj <srijithdamerla2001@gmail.com>